updated carstand

+++ Modulo3/Projecto1/CarStandBusiness/DTO/MarcasDTO.cs
@@ -2,7 +2,7 @@
 {
     public class MarcasDTO
     {
-        public long IDMarca { get; set; }
+        public int IDMarca { get; set; }
         public string Nome { get; set; } = null!;
     }
 }
