+++ Modulo3/Projecto1/CarStandBusiness/Models/Marcas.cs
@@ -2,7 +2,7 @@
 {
     public class Marcas
     {
-        public long IDMarca { get; set; }
+        public int IDMarca { get; set; }
         public string Nome { get; set; } = null!;
     }
 }
