+++ Modulo3/Projecto1/CarStandBusiness/Models/Modelos.cs
@@ -2,7 +2,7 @@
 {
     public class Modelos
     {
-        public long IDModelos { get; set; }
+        public int  IDModelos { get; set; }
         public string Modelo { get; set; } = null!;
     }
 }
